Fix download progress action view layout in RTL mode
The layout_centerInParent attributes caused the parent RelativeLayout to
attempt to expand to the full available width, which didn't make a
difference in LTR mode where it was the rightmost item, but did cause
the view to take over the whole action bar in RTL mode.

This commit changes the container to a FrameLayout with all the children
having their layout_gravity set to center.

MOB-1723 #resolve